requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
Faker==20.1.0
flask==3.0.0
Flask-Cors==4.0.0
flask-sqlalchemy==3.1.1
Flask-SSLify==0.1.5
greenlet==3.0.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
werkzeug==3.0.1
zipp==3.17.0
Flask-JWT-Extended==4.5.3
requests==2.31.0